Microsoft app insights react package sync (#15)
Microsoft application insights react JS package version needs to be updated to 3.0.2 in order to be in sync with applicationinsights-web npm package. Microsoft app insights team has upgraded the package to be always in sync for further releases as mentioned https://github.com/microsoft/ApplicationInsights-JS/issues/1346

Based on the appinsights teams recommendation that applicationinsights-web should be 2.5.7^ to work with applicaitoninsights-react-js packages.